Renames package from main to waitfor.
