+++ cmd/juju/waitfor/application_test.go
@@ -1,7 +1,7 @@
 // Copyright 2020 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package main
+package waitfor
 
 import (
 	"github.com/juju/juju/cmd/juju/waitfor/query"
